clutter-frame-clock: Schedule triple-buffering without presentation times
If triple buffering is enabled then in the DISPATCHED_ONE state a new frame
is allowed to be scheduled before the previous one completes presentation.
But in this state we shouldn't use `calculate_next_update_time_us` because
it would give a `next_update_time_us` that's too old (a repeat), causing
premature rendering and high latency. Instead we calculate the next update
time as precisely one frame interval after the last. Assuming the system
can ever catch up to full frame rate, this avoids the high latency of
classical triple buffering (from the previous commit).

Note that we use the last intended update time and not the last actual
dispatch time, to avoid phase drift from event loop delays.