onscreen/native: Clear only next_fb in page_flip_feedback_discarded
When a flip is discarded it is `next_fb` that failed. However by calling
`meta_onscreen_native_swap_drm_fb` we would end up clearing `current_fb`
that the kernel is still scanning out, and keeping a reference to
`next_fb`. That was the opposite of what we want.

What we want when a `next_fb` flip fails is to keep `current_fb` because
that's what's still scanning out (probably), and to clear the reference
to `next_fb` that's being discarded.