Mark several additional tests as flaky
Most of these have never been reliable on Debian buildds, especially
on ARM, so having them fail is not a regression. They usually succeed
after a few retries, but that's time-consuming and unsuitable to be a
QA gate.

The input-capture test frequently hangs until it times out, so put that
one in a separate suite so that we don't need to run it at all.

Signed-off-by: Simon McVittie <smcv@debian.org>
Forwarded: no