debian: Install gdctl in mutter-common-bin
gdctl is a tool to show and set the gnome display configuration and it's
not mutter-specific, but it works for any-mutter based compositor including
gnome-shell, so it's something that we should not provide in the mutter
package (that only contains the mutter basic compositor implementation).

As per this, move the dependencies to mutter-common-bin and make
python3-argcomplete a required dependency for it (since the tools needs
it).